JavaScript Date Object
This project uses the default Date Object to extract the information
that one wants to display. In this case I extracted month, day and year.
Month is displayed starting from 0 to 11. I just created an array with
the string values of the month. This makes the actual month appear by
name when you string everything together instead of seeing numbers for
month.